backend/requirements.txt
Flask
Flask-CORS
pymongo
python-dotenv
openai
sentence-transformers
numpy
scikit-learn